### 2021-08-26 #128 @whedon
[PRE REVIEW]: Data Carpentry for Biologists: A semester long Data Carpentry course using ecological and other biological examples
**Submitting author:** @ethanwhite (<a href="http://orcid.org/0000-0001-6728-7745">Ethan P. White</a>)
**Repository:** <a href="https://github.com/datacarpentry/semester-biology/" target ="_blank">https://github.com/datacarpentry/semester-biology/</a>
**Version:** v4.1.0
**Managing EiC:** Jordan Gorzalski

Due to the challenges of the COVID-19 pandemic, JOSE is currently operating in a "reduced service mode". You can read more about what that means in [our blog post](https://blog.joss.theoj.org/2020/05/reopening-joss).
