### 2021-08-27 a0b48b3
Remove fancy quotes from paper
