+++ paper.bib
@@ -12,9 +12,9 @@
   volume={5},
   number={6},
   pages={66--70},
-  year={2015}
+  year={2015},
+  doi={10.9790/7388-05616670}
 }
-
 
 @article{barone2017,
   title={Unmet needs for analyzing biological big data: A survey of 704 NSF principal investigators},
@@ -24,7 +24,8 @@
   number={10},
   pages={e1005755},
   year={2017},
-  publisher={Public Library of Science San Francisco, CA USA}
+  publisher={Public Library of Science San Francisco, CA USA},
+  doi={10.1371/journal.pcbi.1005755}
 }
 
 @article{embl2013,
@@ -42,7 +43,8 @@
   number={6},
   pages={546--557},
   year={2017},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
+  doi={10.1093/biosci/bix025}
 }
 
 @article{jones2006,
@@ -52,7 +54,8 @@
   volume={37},
   pages={519--544},
   year={2006},
-  publisher={Annual Reviews}
+  publisher={Annual Reviews},
+  doi={10.1146/annurev.ecolsys.37.091305.110031}
 }
 
 @article{rosenshine1987,
@@ -63,7 +66,8 @@
   number={3},
   pages={34--36},
   year={1987},
-  publisher={Sage Publications Sage CA: Thousand Oaks, CA}
+  publisher={Sage Publications Sage CA: Thousand Oaks, CA},
+  doi={10.1177/002248718703800308}
 }
 
 @article{teal2015,
@@ -72,7 +76,8 @@
   journal={International Journal of Digital Curation},
   volume={10},
   pages={135--143},
-  year={2015}
+  year={2015},
+  doi={10.2218/ijdc.v10i1.351}
 }
 
 @article{williams2019,
@@ -83,5 +88,6 @@
   number={11},
   pages={e0224288},
   year={2019},
-  publisher={Public Library of Science San Francisco, CA USA}
+  publisher={Public Library of Science San Francisco, CA USA},
+  doi={10.1371/journal.pone.0224288}
 }
