+++ paper.bib
@@ -12,8 +12,7 @@
   volume={5},
   number={6},
   pages={66--70},
-  year={2015},
-  doi={10.9790/7388-05616670}
+  year={2015}
 }
 
 @article{barone2017,
