### 2021-08-30 #128 @ShanEllis
Thanks @ethanwhite! Looks ready to start the review process.

Hi @jsta and @jhollist, I have you both as interested reviewers for JOSE, and it appears you are each a good fit for this submission. Would you like to contribute a review for this submission? The title is "Data Carpentry for Biologists: A semester long Data Carpentry course using ecological and other biological examples". Please let me know if you're available and willing. Thanks!
