### 2021-09-14 #128 @pschloss
Hi - thanks for reaching out. Unfortunately, I've been pretty swamped with
other reviews and projects that I need to tend to.

Pat

On Mon, Sep 13, 2021 at 6:19 PM Shannon Ellis ***@***.***>
wrote:

> Hi @pschloss <https://github.com/pschloss>, given your interest in
> reviewing for JOSE (and previous review - thank you!), I wanted to reach
> out to see if you'd be interested in contributing a review for this
> submission? The title is "Data Carpentry for Biologists: A semester long
> Data Carpentry course using ecological and other biological examples".
> Please let me know if you're available and willing in the next couple of
> days. Thanks!
>
