### 2021-09-16 #128 @ShanEllis
Hi @fmichonneau! I have you down as an interested reviewer for JOSE, and I think we have a recent submission that's a really good fit for you. I wanted to reach out to see if you'd be interested in contributing a review for this submission? The title is "Data Carpentry for Biologists: A semester long Data Carpentry course using ecological and other biological examples". Please let me know if you're available and willing in the next couple of days. Thanks!
