### 2021-09-21 #128 @fmichonneau
Hi @ShanEllis Thanks for the invitation but I think my affiliation with The Carpentries creates a conflict of interest for this submission.
