### 2021-10-01 #128 @ShanEllis
Hi @karthik! I'm sure you're super busy (which is why I've held off thus far on bugging you for this), but I've been struggling to pin down reviewers. **Any chance you have time for a JOSE review?** The title is "Data Carpentry for Biologists: A semester long Data Carpentry course using ecological and other biological examples". I believe you've reviewed for JOSE before, so you know that the papers are not very long and the review process is open and happens entirely on GitHub. Let me know if you'll have time to complete this review!
