### 2021-10-01 #128 @labarba
I noticed on the paper it says these materials have been used in a number of university and college courses. Why not find someone among the instructors of those courses to act as reviewers? They would already have read the material and had direct experience teaching with them, perhaps finding areas that need improvement. I don't see that there would be any reason to think adopting the materials raises conflict-of-issue concerns, if the adopter did not collaborate with the original authors... What do you think, @ethanwhite ?
