### 2021-10-01 #128 @ShanEllis
That's a good idea, @labarba. I wasn't sure if there was a COI issue there, but if you and @ethanwhite think that's a  possibility, I'll go ahead with that!

Also, we do have one reviewer on board now! I'll add @smwindecker to the official review issue once we find our second (unless you say otherwise, @labarba!)
