### 2021-10-01 #128 @labarba
> I'll add @smwindecker to the official review issue once we find our second (unless you say otherwise/to start it before then, @labarba!)

Yes, you add the reviewers here, in the Pre-review issue. Go ahead and run `@whedon assign <user> as reviewer` here for the first reviewer, then `@whedon add <user> as reviewer` for the second, then `@whedon start review` 😄
