### 2021-10-01 #128 @ShanEllis
> I agree that folks who have used this material in their courses could be good reviewers. Of the courses on that list the only folks I'm aware of who have conflicts of interest are the University of Florida folks (institutional conflicts + one of them is an author on the paper).
>
> One person on that list that comes to mind as a possibility is catherinehulshof.

Excellent. Thank you, @ethanwhite. I will reach out to to this end, keeping the exceptions you noted in mind. I appreciate it!
