### 2021-10-01 #128 @ShanEllis
Hi @catherinehulshof! Reaching out to see if you're interested and able to review a submission to JOSE (Journal of Open Source Education) titled "Data Carpentry for Biologists: A semester long Data Carpentry course using ecological and other biological examples".  I believe you are familiar with these materials, making you a good fit for the review and hopefully making the review process shorter for you. The review process for JOSE is open and happens on GitHub. If you're interested and able, please let me know!
