### 2021-10-02 #128 @catherinehulshof
Dear Shannon,

I would be interested and am familiar with the materials after learning
more about the review process. What are the review criteria/format, etc?

Cheers,

Catherine

______________________________________________
Catherine M. Hulshof, PhD (she/her: why I use pronouns
<https://www.mypronouns.org/what-and-why>)
Environmental Data Science
Department of Biology
Virginia Commonwealth University
Richmond, VA, 23284
https://www.biodiversityresearchlab.com

On Fri, Oct 1, 2021 at 7:40 PM Shannon Ellis ***@***.***>
wrote:

> Hi @catherinehulshof <https://github.com/catherinehulshof>! Reaching out
> to see if you're interested and able to review a submission to JOSE
> (Journal of Open Source Education) titled "Data Carpentry for Biologists: A
> semester long Data Carpentry course using ecological and other biological
> examples". I believe you are familiar with these materials, making you a
> good fit for the review and hopefully making the review process shorter for
> you. The review process for JOSE is open and happens on GitHub. If you're
> interested and able, please let me know!
>
