### 2021-10-04 #128 @catherinehulshof
Hi Shannon,

I'd be happy to provide a review within the next several weeks.

Catherine

______________________________________________
Catherine M. Hulshof, PhD (she/her: why I use pronouns
<https://www.mypronouns.org/what-and-why>)
Environmental Data Science
Department of Biology
Virginia Commonwealth University
Richmond, VA, 23284
https://www.biodiversityresearchlab.com

On Mon, Oct 4, 2021 at 2:31 PM Shannon Ellis ***@***.***>
wrote:

> @catherinehulshof <https://github.com/catherinehulshof> Oh excellent! To
> explain the process/steps a bit:
>
>    1. Once agreed, I would add you officially as a reviewer.
>    2. After this I would start the review officially (rather than this
>    issue, which is "pre-review") where you and the other reviewer would begin
>    your review.
>    3. Upon starting the review, we would work from a new issue on GitHub.
>    There would be a checklist here (to guide your review) and your review
>    would be included as a comment there.
>    4. Upon receipt of both reviews (ideally within a few weeks, but we
>    understand people are busy), I would review your comments along with the
>    other reviewer and communicate with the authors at that point in time.
>
> To see an example of a review issue (of a paper that was subsequently
> published), have a look here: #115
> <https://github.com/openjournals/jose-reviews/issues/115>
>
> Also, a note that whedon is the bot that helps make this process go as
> smoothly as possible, just so you know what it means whenever you see that
> pop up. Happy to answer any questions you may have if this doesn't clear
> things up!
>
>    - Shannon
>
