### 2021-10-04 #139 @whedon
[REVIEW]: Data Carpentry for Biologists: A semester long Data Carpentry course using ecological and other biological examples
**Submitting author:** @ethanwhite (<a href="http://orcid.org/0000-0001-6728-7745">Ethan P. White</a>)
**Repository:** <a href="https://github.com/datacarpentry/semester-biology/" target ="_blank">https://github.com/datacarpentry/semester-biology/</a>
**Version:** v4.2.0

- [x] #145
