### 2021-10-14 0804728
Expressions and Variable translation
Spanish translation of exercises of expressions and variables topic
