+++ exercises-es/Expressions-and-variables-variable-reassignment-R-es.md
@@ -1,0 +1,10 @@
+Usando una sola variable, convierta una temperatura de 42 grados Fahrenheit
+a Celsius. Para hacer esto, siga los siguientes pasos:
+
+1. Asigne el número 42 a una variable.
+2. Restele 32 y asigne ese valor a la misma variable.
+3. Divida el valor actual por 9 y asigne ese valor a la misma
+    variable.
+4. Luego multiplique el valor actual por 5 y asigne ese valor nuevamente
+    a la misma variable.
+5. Imprima la temperatura en grados Celsius.
