+++ exercises-es/Expressions-and-variables-variable-names-R-es.md
@@ -1,0 +1,11 @@
+En el ejercicio [Más sobre variables] ({{site.baseurl}} / ejercicios-es / Expressions-and-variables-more-variables-R-es) usamos los nombres de las variables `grams` (gramos),` number`(número) y `biomass` (biomasa) para describir la masa individual, número de individuos y biomasa total de algunas ratas de bosque de garganta blanca.
+Pero, ¿son estos nombres de variables la mejor opción?
+Si volviéramos al código para esta asignación en dos semanas, ¿recordariamos a qué se referían estas variables y, por lo tanto, qué estaba pasando en el código?
+
+El nombre de la variable "biomass" (biomasa) también es bastante largo.
+Si tuviéramos que escribirlo muchas veces, sería más rápido escribir `b`.
+También podríamos utilizar alternativas realmente descriptivas como `individual_mass_in_grams`(masa_de_individuo_en_gramos).
+Otra alternativa seria abreviar este nombre u omitir algunas de las palabras para hacer el nombre más corto (por ejemplo, `indiv_mass_g`)(masa_indiv_g).
+
+Piense en buenos nombres de variables y cree una nueva versión de este código con las variables renombradas para que sean más útiles.
+Asegúrese de que su código se ejecute correctamente con los cambios de nombre.
