+++ exercises-es/Expressions-and-variables-more-variables-R-es.md
@@ -1,0 +1,14 @@
+Calcule la biomasa total en gramos para 3 white-throated woodrats
+([* Neotoma albigula *] (https://en.wikipedia.org/wiki/White-throated_woodrat)) y
+luego convierta la biomasa a kilogramos. La biomasa total es tres veces el peso promedio 
+de un solo individuo. Un individuo promedio pesa 250 gramos.
+
+* Agregue una nueva sección a su R script comenzando con un comentario.
+* Cree una variable `grams` (gramos) y asígnele la masa de un solo *Neotoma albigula* (250).
+* Cree una variable `number` (número) y asígnele el número de individuos (3).
+* Cree una variable `biomass` (biomasa) y asígnele un valor multiplicando
+las variables `grams` (gramos) y `number`(número) juntas.
+* Convertir el valor de `biomass`(biomasa) en kilogramos (hay 1000
+gramos en un kilogramo, así que divida entre 1000) y asigne este valor a una nueva
+variable.
+* Imprima la respuesta final en la pantalla.
