+++ exercises-es/Expressions-and-variables-basic-expressions-R-es.md
@@ -1,0 +1,32 @@
+Escriba los siguientes cálculos en el editor de texto.
+
+1. 2 - 10
+2. 3 \* 5
+3. 9 / 2
+4. 5 - 3 \* 2
+5. (5 - 3) \* 2
+6. 4 ^ 2
+7. 8 / 2 ^ 2
+
+Ejecútelos haciendo clic en el botón `Run` en la esquina superior derecha del editor o presione Ctrl + Enter
+(Windows y Linux) o Command+Enter (Mac) para ejecutar e imprimir el código y los resultados en la consola.
+
+Si no hay código resaltado/seleccionado, se ejecutará la línea en la que se encuentra el cursor.
+Si resaltó / seleccionó un bloque de código, se ejecutará ese grupo de líneas.
+
+También puede ejecutar el 'script' completo haciendo clic en la flecha junto a `Source` y seleccionando `Source with Echo` o usando Ctrl + Shift + Enter
+(Windows y Linux) o Command+Shift+Enter (Mac).
+
+Agregue una línea de comentario que diga 'Ejercicio 1' antes del código que corresponde a el ejercicio. Los comentarios en R se agregan con el símbolo '#'.
+Cualquier texto luego de el símbolo '#' en la misma línea se ignora cuando el programa se
+ejecuta. La primera parte de su programa debería verse así:
+
+``` r
+# Exercise 1
+
+N.º 1.1
+2 - 10
+
+N.º 1.2
+3 * 5
+```
