+++ exercises-es/Vectors-nulls-in-vectors-R-es.md
@@ -1,0 +1,19 @@
+---
+layout: exercise
+topic: Vectors
+title: Nulls in Vectors
+language: R
+translation: es
+titulo: Valores nulos en vectores
+---
+Corte y pegue el siguiente vector en su tarea y luego use código para imprimir los valores solicitados en la pantalla.
+Use `na.rm = TRUE` para ignorar los valores nulos.
+
+```
+numbers <- c(7, 6, 22, 5, NA, 42)
+```
+
+1. El número más pequeño en el vector `numbers` 
+2. El número más grande en el vector `numbers` 
+3. El promedio de los números en los `numbers` 
+4. La suma de los valores en el vector `numbers`
