+++ exercises-es/Vectors-basic-vectors-R-es.md
@@ -1,0 +1,22 @@
+---
+layout: exercise
+topic: Vectors
+title: Basic Vectors
+language: R
+translation: es
+titulo: Vectores basicos
+---
+Corte y pegue el siguiente vector en su tarea y luego use `print()` para
+imprimir los valores solicitados en la pantalla.
+
+```
+numbers <- c(5, 2, 26, 8, 16)
+```
+
+1. El número de elementos en el vector `numbers` (usando la función `length`)
+2. El tercer elemento en el vector `numbers` (usando `[]`)
+3. El número más pequeño en el vector `numbers` (usando la función `min`)
+4. El número más grande en el vector `numbers` (usando la función `max`)
+5. El promedio de los números en el vector `numbers` (usando la función `mean`)
+6. El primer, segundo y tercer número en el vector `numbers` (usando `[]`)
+7. La suma de los valores en el vector `numbers` (usando la función `sum`)
