### 2021-10-20 #139 @smwindecker
Hi all,

I commend the authors on a wonderful resource for teaching. We recently began discussions here in Australia about consolidated teaching resources that very much match this tool. I will bring this model to my colleagues with pleasure.

My favourite parts of the work are the clear navigation between pages, the look and content in the schedule page, and the customisability. It strikes me as a great resource for reducing duplication of effort in teaching these foundational skills.

_Stated Dependencies_
- I was only uncertain on one box, the "Installation Instructions", as it asked for clearly stated dependencies. I do not think this requirement is particularly relevant to this work as it does not appear there are any dependencies, besides jekyll if you want to check the built website from your computer before pushing.
- In any case, if there is something I have overlooked the authors could consider a DESCRIPTION file with the software/packages, citation/authors, and other dependencies.

_Instructions for customisation_
- Although the authors have done a great job, and there are lots of evident examples in the wild of this resource being modified, I think there are still some barriers for less github savvy or merely time-poor users to modify all necessary code and set up the site.
- One idea to simplify this process would be to have a guide to new instructors that lists all the places they need to edit, and keep them all together. The instructions currently state a new user would only have to modify the _config.yml if they do not wish to/need to modify the subject matter, but the syllabus page, for example, has the University of Florida canvas link, and the audience.md file has the original course email. I may have missed other examples of places that would need to be modified.
- I confess I set up my _config.yml on my fork, committed and pushed, but have not gotten anything other than the 404 warning to my github.io link. These things sometimes take a bit of time though..

_Minor notes_
- as mentioned in the previous review, there are a few dead links. Another example was the syllabus.md page 'info-graphic'
- I would also like to second another comment in the previous review regarding the "talking heads" in the videos. With respect, I agree that it would possibly improve uptake without the speaker, or with more diverse presenters. Having different faces, or no faces, may also make it easier for instructors to assign the videos in their classes without feeling that they are deferring instruction to an entire set of another instructor's course lectures.

Congratulations again on a well executed resource. I hope it is widely shared, built upon, and used. I recommend strongly for publication.
