### 2021-11-03 datacarpentry/semester-biology#915 @ethanwhite
Implement broken link checking
A number of broken links were reported in https://github.com/openjournals/jose-reviews/issues/139. We should implement link checking using GH Actions and fix the broken links identified.
