### 2021-11-03 datacarpentry/semester-biology#916 @ethanwhite
Improve instructions for customization
Create a functioning fork of the course is still challenging (see https://github.com/openjournals/jose-reviews/issues/139). We should add videos on how to fork and modify the course as well as do a full review of the current written documentation and look for places to improve.
