### 2021-11-03 datacarpentry/semester-biology#917 @ethanwhite
Add dependencies for local site building
Jekyll and associated gems. See https://github.com/openjournals/jose-reviews/issues/139
