### 2021-11-03 datacarpentry/semester-biology#918 @ethanwhite
Fixed 'Expected outputs for...' on non-exercise/assignment pages
A number of non-exercise/assignment pages have 'Expected outputs for...' at the bottom of them, which should only be on exercise/assignment pages to provide links to output solutions.
