### 2021-11-03 datacarpentry/semester-biology#919 @ethanwhite
Improve representation in videos or remove instructor video
Since all of the videos were made for the Fall of 2020 run of my course (thanks COVID!) the only person represented in the video material is a cis het white man. This perpetuates existing inequities and lack of representation in CS and computational biology. A big thanks to @catherinehulshof & @smwindecker for highlighting this in https://github.com/openjournals/jose-reviews/issues/139.

There are two approaches to addressing this issue:

1. Engage other contributors in developing replacement videos
2. Remove the talking heads from the videos (which may also make them easier to engage with for some viewers)

Unfortunately the talking heads can't be removed from the existing videos, so either of these approaches represents an investment of ~40 hours/lesson.

If anyone is interested in working on this I am happy to provide extensive support including supporting grant proposals to cover salary for doing the work.
