### 2021-11-03 #139 @ethanwhite
Thanks so much for these excellent reviews @catherinehulshof & @smwindecker!

I particularly appreciate your efforts to highlight areas in which we can improve with respect to representation. It's something I've been working on from an in-classroom perspective with UF instructional designers and definitely an issue in the online environment. Unfortunately given the way the videos were recorded the talking heads can't be removed from the existing recordings and re-recording, editing, & captioning would be ~400 hours worth of work which isn't feasible in the foreseeable future. I've opened a long-term issue with the hopes of addressing this important point (https://github.com/datacarpentry/semester-biology/issues/919) and will look into funding sources (including Broader Impacts on upcoming grant proposals) as a way of supporting this work financially.

I've also opened issues to be dealt with before publication for:
- [ ] Broken links: https://github.com/datacarpentry/semester-biology/issues/915
- [ ] Incorrect 'Expected outputs for..." lines: https://github.com/datacarpentry/semester-biology/issues/918
- [ ] An adding dependencies for local site builds: https://github.com/datacarpentry/semester-biology/issues/917

Finally I've also opened an issue for general improvements to the instructions for customization including the addition of videos and improvements to written instructions: https://github.com/datacarpentry/semester-biology/issues/916. Based on your feedback I'm currently considering this more of a long-term issue, but if I misunderstood just let me know.
