### 2021-11-04 datacarpentry/semester-biology#915 @andrewmarx
There's a 'html-proofer' Ruby gem that might work for this. Here is a short tutorial explaining how to implement it as a github action: https://clementbm.github.io/github%20action/jekyll/link%20checking/2020/05/31/automatically-validate-links-on-jekyll-website.html

I'm not familiar enough with Jekyll/Ruby/Github actions to figure out what needs to be adjusted from that tutorial for the course website, but I could maybe do it in the next month or two when I have time if no one else decides to tackle it before then.
