### 2021-11-29 #139 @ShanEllis
Hi @ethanwhite - wanted to check in on progress here! I know we're into the holidays at this point, and I'm sure you're quite busy, but I wanted to see if it'd be reasonable to address these reviewer comments in the coming weeks (maybe, by December 17th?)

I'm definitely here to work with you if that doesn't work; I just know I do better when I have a deadline in my calendar.
