+++ exercises/Neon-canopy-height-from-space-R.md
@@ -5,7 +5,7 @@ title: Canopy Height from Space
 language: R
 ---
 
-The [National Ecological Observatory Network](http://www.neonscience.org) has invested in [high-resolution airborne imaging](http://www.neonscience.org/data-resources/get-data/airborne-data) of their field sites. 
+The [National Ecological Observatory Network](http://www.neonscience.org) has invested in [high-resolution airborne imaging](https://www.neonscience.org/data-collection/airborne-remote-sensing) of their field sites. 
 Elevation models generated from [LiDAR](http://neondataskills.org/self-paced-tutorial/1_About-LiDAR-Data-Light-Detection-and-Ranging_Activity1/) can be used to map the topography and vegetation structure at the sites.
 
 Check to see if there is a `data` directory in your workspace with an `SJER` subdirectory in it.
