+++ exercises/Neon-neon-raster-R.md
@@ -7,7 +7,7 @@ language: R
 
 Jane just got a job working to develop data products for [NEON](http://www.neonscience.org/). 
 She's got to familiarize herself with working with LiDAR data and seeks out the 
-great 'Work with Data' [tutorial for raster data](http://neondataskills.org/lidar-data/lidar-data-rasters-in-R/). 
+great 'Work with Data' [tutorial for raster data](https://www.neonscience.org/resources/learning-hub/workshops/work-lidar-derived-rasters-r). 
 Help Jane take the next step and develop a few additional products from the 
 [LiDAR Dataset](http://www.neonhighered.org/Data/LidarActivity/CHM_InSitu_Data.zip)
 
