### 2021-11-29 a781bff
Update old link to Hadley's style guide
