### 2021-11-29 671692c
Replace links to old Stat545 site
