+++ readings/R-advanced.md
@@ -12,7 +12,7 @@ language: R
 
 * Readings
 
-  * [Regular Expressions](http://stat545-ubc.github.io/block022_regular-expression.html)
+  * [Regular Expressions](https://bookdown.org/rdpeng/rprogdatascience/regular-expressions.html)
   * [Debugging](http://adv-r.had.co.nz/Exceptions-Debugging.html)
   * [Testing Functions](http://swcarpentry.github.io/r-novice-inflammation/02-func-R/)
   * [Improve Your Programming Vocabulary](http://adv-r.had.co.nz/Vocabulary.html)
