### 2021-11-29 5754d81
Fix flipped classroom info graphic links
