### 2021-11-29 8322c9c
Link fixes (#925)
* Fix links based on GitHub and branch changes raw/master is now blob/main

* Update NEON airborne data link

* Update old neondataskills.org links to neonscience.org

* Update UF specific assessment plan link

* Update old link to Hadley's style guide

* Replace links to old Stat545 site

* Replace broken link to information on rowwise in dplyr

* Update moved w3schools link

* Fix flipped classroom info graphic links

* Fix broken link to Anaconda

* Best guess replacement for broken link to data source. This exercise is currently not in use anywhere.

* Fix broken link to 
[…]
