+++ readings/SQL-queries.md
@@ -8,4 +8,4 @@ language: SQL
 * [Databases Intro](https://www.youtube.com/watch?v=kaKa6N9lEG8)
 * [Basic Queries](http://www.datacarpentry.org/sql-ecology-lesson/01-sql-basic-queries) - [Selecting](https://www.youtube.com/watch?v=Gua3FpRzLdQ), [Filtering](https://www.youtube.com/watch?v=c3hoWxukrPE), [Sorting](https://www.youtube.com/watch?v=rNwwdijxxKs), [Nulls](https://www.youtube.com/watch?v=KLugfNdGNFw)
 * Aggregation - [Video](https://www.youtube.com/watch?v=ZjuL-pfkUOA), [Reading](http://www.datacarpentry.org/sql-ecology-lesson/02-sql-aggregation)
-* Joins - [Video](https://www.youtube.com/watch?v=79EBoVPUzkE), [Reading](http://www.datacarpentry.org/sql-ecology-lesson/03-sql-joins-aliases)
+* Joins - [Video](https://www.youtube.com/watch?v=79EBoVPUzkE), [Reading](https://datacarpentry.org/sql-ecology-lesson/03-sql-joins/index.html)
