### 2021-12-13 datacarpentry/semester-biology#915 @ethanwhite
@andrewmarx - definitely feel encouraged to give this a go if you have the time/energy. I thought I had a quick and easy solution but it failed. I used `linkchecker` to do a manual check and have fixed all of the issues, but would love to have this running on builds and a weekly cron job to prevent the build up of broken links.
