### 2021-12-15 258ff14
Docs pages should be pages not exercises
This will remove the incorrect 'Expected outputs for' language from the
bottom of the pages.

Fixes #918.
