+++ docs/course/course-delivery.md
@@ -1,5 +1,5 @@
 ---
-layout: exercise
+layout: page
 topic: Help
 title: Teaching Strategies and Course Delivery
 ---
