### 2021-12-15 datacarpentry/semester-biology#915 @ethanwhite
Removing `jose-publication` milestone since the broken links themselves have been fixed satisfying the requirements for https://github.com/openjournals/jose-reviews/issues/139
