### 2021-12-15 #139 @ethanwhite
Thanks for the ping @ShanEllis - I was hoping to have this back to you before the break, but I'm juggling enough end-of-year deadlines that it's not quite going to happen. I should be able to get it back in the first week of January.
