### 2021-12-16 #139 @ShanEllis
Totally understand that @ethanwhite! Sounds great re: first week of January. Enjoy break!
