+++ .github/workflows/checklinks.yml
@@ -15,28 +15,18 @@ jobs:
   checklinks:
     name: Linux
     runs-on: ubuntu-latest
-    strategy:
-      fail-fast: false
     steps:
       - uses: actions/checkout@v2
-      - name: Ruby
-        uses: actions/setup-ruby@v1
+      - uses: actions/setup-ruby@v1
         with:
           ruby-version: 2.6.x
-      - name: Setup Rubygems, Bundler, jekyll
-        run: | 
-          gem update --system --no-document
-          gem update bundler --no-document
-          gem install jekyll bundler
+      - run: |
           bundle install
-      - name: Build jekyll website with drafts
-        run: bundle exec jekyll build --drafts
-      - name: Check for broken links
-        run: |
-          bundle exec htmlproofer --log-level :debug ./_site &> links.log
-        continue-on-error: true
-      - name: Archive log links
-        uses: actions/upload-artifact@v1
+          gem install html-proofer
+      - run: bundle exec jekyll build
+      - uses: nick-invision/retry@v2
         with:
-          name: links-check.log
-          path: links.log
+          max_attempts: 3
+          timeout_minutes: 10
+          retry_on: error
+          command: htmlproofer ./_site
