### 2021-12-17 1f4f73d
Update checklinks.yml
https://github.com/gjtorikian/html-proofer#using-with-jekyll
