+++ .github/workflows/checklinks.yml
@@ -29,4 +29,4 @@ jobs:
           max_attempts: 3
           timeout_minutes: 10
           retry_on: error
-          command: htmlproofer ./_site
+          command: htmlproofer --assume-extension ./_site
