### 2021-12-17 b2abcef
Update checklinks.yml
Adjust for Jekyll's baseurl in the config file

Reduce attempts/timeout for now
