### 2021-12-19 342ed57
Update checklinks.yml
Remove retry code

Set to continue on error
