+++ .github/workflows/checklinks.yml
@@ -24,9 +24,5 @@ jobs:
           bundle install
           gem install html-proofer
       - run: bundle exec jekyll build
-      - uses: nick-invision/retry@v2
-        with:
-          max_attempts: 1
-          timeout_minutes: 5
-          retry_on: error
-          command: htmlproofer --assume-extension ./_site --url-swap '^/semester-biology/:/'
+      - run: htmlproofer --assume-extension ./_site --url-swap '^/semester-biology/:/'
+        continue-on-error: true
