### 2022-01-05 datacarpentry/semester-biology#915 @ethanwhite
Completed by @andrewmarx in #935
