### 2022-01-05 #139 @ethanwhite
I have now completed the revision. Thanks again to @catherinehulshof & @smwindecker for both their detailed fixes and their broader long-term suggestions! All of the issues listed in my comment above have been addressed and closed and a new release made: https://github.com/datacarpentry/semester-biology/releases/tag/v4.2.0. This release is available on Zenodo at https://doi.org/10.5281/zenodo.5821925.

@ShanEllis - let me know if there's anything else you need on my end.
