### 2022-01-12 #139 @ShanEllis
Thank you so much @ethanwhite! (Sorry for delay on my end.)

At this point, I'd like to give @catherinehulshof and @smwindecker a chance to look over the authors' responses and changes made to see if they're satisfied at this point and if they recommend publication?

Also, I'll begin now, and will continue once we have responses from both reviewers, to do final checks and make any further small suggestions (if needed) prior to final checks/sending this paper off officially!
