+++ Aggregation-count-Access-es.md
@@ -1,0 +1,8 @@
+---
+layout: exercise
+topic: Aggregation
+title: Count
+language: Access]
+---
+Escribe una consulta (query) que devuelva el número de todos los individuos (de toas las especies) que han sido capturados en cada año, ordenados cronológicamente. Incluya el año
+en la salida. Guárdelo como "Abundancia total por año".
