### 2022-02-02 2e923b1
Delete Aggregation-count-SQL-es.md
Added to wrong folder
