### 2022-02-02 a3e0967
Delete Aggregation-sum-Access-es.md
Added to wrong folder
