+++ Aggregation-sum-Access-es.md
@@ -1,12 +1,0 @@
----
-layout: exercise
-topic: Aggregation
-title: Sum
-language: Access
----
-
-Escribe una consulta (query) que devuelva el número de individuos de cada especie que han sido
-capturados en cada año y la biomasa total de esos individuos (la
-suma de la columna wgt). Incluya el año y la identificación de la especie en la salida.
-Ordene el resultado cronológicamente por año y luego alfabéticamente por
-especies. Guarde como `Mass-Abundance Data`.
