### 2022-02-02 edd286a
Delete Aggregation-sum-SQL-es.md
Added to wrong folder
