+++ Aggregation-sum-SQL-es.md
@@ -1,13 +1,0 @@
----
-layout: exercise
-topic: Aggregation
-title: SUM
-language: SQL
----
-
-Escriba una consulta (query) que devuelva el número de individuos de cada especie que han sido
-capturados en cada año (`total_abundance`) y la biomasa total, `total_biomass`, de aquellos
-individuos (* la suma de `weight` *). Las unidades de biomasa deben estar en
-kilogramos. Incluya el `year` y el `species_id` en la salida. Ordenar el resultado
-cronológicamente por año y luego alfabéticamente por especie. Guardar como
-`mass_abundance_data`.
