### 2022-02-02 4544999
Tidy Data
