+++ exercises-es/Tidy-data-improving-messy-data-SQL-es.md
@@ -1,0 +1,21 @@
+---
+layout: exercise
+topic: Tidy Data
+title: Improving Messy Data
+language: SQL
+---
+Muchos datos reales no están muy ordenados, principalmente porque a la mayoría de los científicos no se les enseña a estructurar sus datos de una manera que sea fácil de analizar.
+
+
+[Download an untidy version](https://ndownloader.figshare.com/files/24469424),
+estos son algunos de los datos desordenados del Proyecto Portal, que incluye información sobre el sitio, fecha,identificación de especies, peso y parcela de muestreo (dentro de la area de estudio) para algunos pequeños mamíferos.
+
+Piense en lo que se podría mejorar de estos datos y escriba las respuestas a las siguientes preguntas:
+
+1. Describa cinco cosas sobre estos datos que no están ordenadas (tidy) y cómo podría
+     solucionar cada uno de esos problemas.
+
+2. ¿Se podrían importar fácilmente estos datos a un lenguaje de programación o a un
+     base de datos en su forma actual?
+
+3. ¿Cree que es una buena idea ingresar los datos de esta manera y luego limpiarlos más tarde, o seria mejor tener una buena estructura de datos para el análisis  desde el momento en que se ingresan los datos? ¿Por qué?
