### 2022-02-02 1e7aa5e
Delete Loops-size-estimates-by-name-apply-R-es.md
Uploaded to wrong folder
