### 2022-02-02 974c385
Delete Qaqc-data-entry-validation-in-excel-SQL-es.md
Uploaded to wrong folder
