### 2022-02-02 0153b8d
Delete Making-choices-size-estimates-by-name-R-es.md
Uploaded to wrong folder
