### 2022-02-03 e585aa6
Delete Loops-size-estimates-by-name-loop-R-es.md
Uploaded to wrong folder
