### 2022-02-03 00fa511
Delete Loops-size-estimates-vectorized-R-es.md
Uploaded to wrong folder
