### 2022-02-03 e3fc52b
Delete Making-choices-load-or-download-file-R-es.md
Uploaded to wrong folder
