### 2022-02-03 9cc113b
Delete Loops-size-estimates-with-maximum-R-es.md
Uploaded to wrong folder
