### 2022-02-03 2d224a0
Delete Making-choices-choice-operators-R-es.md
Uploaded to wrong folder
