+++ Making-choices-choice-operators-R-es.md
@@ -1,31 +1,0 @@
----
-layout: exercise
-topic: Making Choices
-title: Choice Operators
-language: R
----
-Cree las siguientes variables.
-
-```
-w <- 10.2
-x <- 1.3
-y <- 2.8
-z <- 17.5
-colors <- c("red", "blue", "green")
-masses <- c(45.2, 36.1, 27.8, 81.6, 42.4)
-dna1 <- "attattaggaccaca"
-dna2 <- "attattaggaacaca"
-```
-
-Úselas para imprimir si las siguientes premisas son "TRUE" o "FALSE".
-
-1. `w` es mayor que 10
-2. `"green"` está en `colors`
-3. `x` es mayor que `y`
-4. Cada valor en `mass` es mayor que 40.
-5. 2 * `x` + 0.2 es igual a `y`
-6. `dna1` es lo mismo que `dna2`
-7. `dna1` no es lo mismo que `dna2`
-8. `w` es mayor que `x`, o `y` es mayor que `z`
-9. 'x' por 'w' está entre 13.2 y 13.5
-10. Cada masa en `mass` está entre 30 y 50.
