+++ Capstone-length-of-floods-R-es.md
@@ -1,12 +1,0 @@
----
-layout: exercise
-topic: Capstone
-title: Length of Floods
-language: R
----
-
-Usted interesado en estudiar el efecto del tiempo y la duración de los pequeños
-eventos de inundación a escala en un ecosistema. Para hacer esto, debe determinar cuándo se producen las inundaciones y su duración según los datos de la estación de aforo.
-
-Descargar el
-[stream guage data for USGS stream gauge site 02236000]({{ site.baseurl }}/data/st_johns_stream_data.csv) de el río St. Johns en Florida. Encuentre las partes continuas de las series de tiempo donde el nivel de la corriente está por encima del umbral de inundación de 2.26 pies y almacene la información sobre la fecha de inicio y la duración de cada inundación en un marco de datos.
