### 2022-02-04 b540115
Typo fix in file
Corrected vectos to vectors
