+++ exercises-es/Dplyr-shrub-volume-aggregation-R-es 2.md
@@ -1,3 +1,9 @@
+---
+layout: exercise
+topic: dplyr
+title: Shrub Volume Aggregation
+language: R
+---
 Este ejercicio es una continuación de [Shrub Volume Data Basics]({{ site.baseurl }}/exercises/Dplyr-shrub-volume-data-basics-R).
 
 La Dra. Granger quiere un resumen de datos tanto para las plantas en su sitio de estudio como para sus experimentos.Verifique si el archivo `shrub-volume-data.csv` está en su espacio de trabajo (es posible que su instructor ya lo haya agregado). 
