### 2022-03-31 #139 @labarba
Great. @ShanEllis : Your reviewers are happy now. Go ahead and run `@whedon generate pdf` and give it one last editorial pass, with any final copy editing suggestions to the author. Once it looks good, ask the authors to issue a tagged release, then archive their repo in Zenodo or similar, and run the commands to set the version and archive DOI here. Then run `@whedon recommend-accept`. Let me know if you have any questions.
