### 2022-04-01 ab6a4d3
Add Hatch projects to acknowledgements
