+++ exercises/Good-style-format-the-code-R.md
@@ -7,7 +7,7 @@ language: R
 
 Programming computers to do cool science is a major advantage for modern scientists. But, developing research programs that are transparent, collaborative and reproducible is a major advantage for science. All data scientists can contribute to this goal by writing their code in easily readable, well described, and well commented scripts.
 
-[This script]({{ site.baseurl }}/code/bad-style-R.R), which evaluates [tree composition in a Michigan swamp](http://www.esapubs.org/archive/ecol/E090/251/default.htm), doesn't follow 
+[This script]({{ site.baseurl }}/code/bad-style-R.R), which evaluates [tree composition in a Michigan swamp](http://www.esapubs.org/archive/ecol/E090/251/default.html), doesn't follow 
 good style. *It works, but it's difficult to understand exactly what it is doing.* 
 
 - Download [the script]({{ site.baseurl }}/code/bad-style-R.R). 
