+++ exercises-es/Aggregation-count-SQL-es.md
@@ -1,0 +1,11 @@
+---
+layout: exercise
+topic: Aggregation
+title: COUNT
+language: SQL
+---
+
+Escriba una consulta (query) que devuelva el número de individuos identificados por especies en cada año (es decir, cuente la columna `species_id`).
+Nombre la columna de conteo "total_abundance" y ordénela cronológicamente.
+Incluya el año en la salida y guárdelo como `total_abundance_by_year`.
+Solo debe haber un valor para cada año, ya que este es un conteo de los individuos de todas las especies en ese año.
