### 2022-04-13 #139 @labarba
@ethanwhite 👋  — We ask that authors change the metadata of the Zenodo deposit so the title matches the JOSE paper. Could you do that at https://doi.org/10.5281/zenodo.5821925 ?
